Auth: Extract useStartUserLogin to be shared
