+++ src/app/regionen/[regionSlug]/_components/DownloadModal/DownloadModal.tsx
@@ -2,9 +2,9 @@ import { ArrowDownTrayIcon } from '@heroicons/react/24/outline'
 import Cookies from 'js-cookie'
 import { useRouter } from 'next/navigation'
 import { IconModal } from 'src/app/_components/Modal/IconModal'
-import { cookieName } from 'src/app/_components/layouts/Header/User/cookieName'
 import { linkStyles } from 'src/app/_components/links/styles'
 import { useHasPermissions } from 'src/app/_hooks/useHasPermissions'
+import { cookieName } from 'src/users/hooks/cookieName'
 import { useRegion } from '../regionUtils/useRegion'
 import { DownloadModalDownloadList } from './DownloadModalDownloadList'
 import { DownloadModalUpdateDate } from './DownloadModalUpdateDate'
