+++ src/users/hooks/useStartUserLogin.ts
@@ -1,0 +1,12 @@
+import { useRouter } from 'next/navigation'
+import Cookies from 'js-cookie'
+import { cookieName } from './cookieName'
+
+export const useStartUserLogin = () => {
+  const router = useRouter()
+
+  return () => {
+    Cookies.set(cookieName, `${location.pathname}${location.search}`)
+    void router.push('/api/auth/osm/login')
+  }
+}
